engine: refactor ExecutionContext into its own file
